requirements.txt
torch
torchvision
torchaudio
open-clip-torch
numpy
pillow
opencv-python
requests
transformers
ultralytics
